refactor: :art: rename Header to Nav, refactor layout styles for max content width, wipe Projects
