+++ src/components/layout.tsx
@@ -3,7 +3,7 @@ import styled, { ThemeProvider } from 'styled-components'
 
 import theme from '../theme'
 import GlobalStyles from '../theme/globalStyles'
-import Header from './Header'
+import Nav from './Nav'
 import Footer from './Footer'
 import { useLocalStorage } from '../utils/useLocalStorage'
 
@@ -29,13 +29,11 @@ const Layout: FC<Props> = ({ location, children }) => {
 		<ThemeProvider theme={newTheme}>
 			<GlobalStyles />
 			<Wrapper>
-				<StyledContent>
-					<Header
-						location={location}
-						toggleTheme={() => setDarkMode(!isDarkMode)}
-					/>
-					<main>{children}</main>
-				</StyledContent>
+				<Nav
+					location={location}
+					toggleTheme={() => setDarkMode(!isDarkMode)}
+				/>
+				<main>{children}</main>
 				<Footer />
 			</Wrapper>
 		</ThemeProvider>
@@ -43,14 +41,15 @@ const Layout: FC<Props> = ({ location, children }) => {
 }
 
 const Wrapper = styled.div`
+	max-width: 60rem;
 	min-height: 100vh;
 	display: flex;
 	flex-direction: column;
-`
-
-const StyledContent = styled.div`
-	width: 100%;
-	flex: 1 0 auto;
+	padding: ${({ theme }) => theme.spacing(0, 5)};
+	margin: 0 auto;
+	${({ theme }) => theme.breakpoints.down('xs')} {
+		padding: ${({ theme }) => theme.spacing(0, 2)};
+	}
 `
 
 export default Layout
